// src/app/shared/generated/model/agentStatus.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum AgentStatus {
    Draft = 'DRAFT',
    Live = 'LIVE',
    Deprecated = 'DEPRECATED'
}